Update the example
